+++ cpp/include/multigrid/detail/coarsest.hpp
@@ -7,19 +7,19 @@
 namespace top3d { namespace mg {
 
 void EleMod_CompactToFull_Finest(const Problem& pb,
-                                 const std::vector<double>& eleModCompact,
-                                 std::vector<double>& eleModFull);
+                                 const std::vector<float>& eleModCompact,
+                                 std::vector<float>& eleModFull);
 
 void MG_AssembleCoarsestDenseK_Galerkin(const Problem& pb,
                                         const MGHierarchy& H,
-                                        const std::vector<double>& eleModFineFull,
+                                        const std::vector<float>& eleModFineFull,
                                         const std::vector<uint8_t>& fineFixedDofMask,
-                                        std::vector<double>& Kc);
+                                        std::vector<float>& Kc);
 
-bool chol_spd_inplace(std::vector<double>& A, int N);
-void chol_solve_lower(const std::vector<double>& L,
-                      const std::vector<double>& b,
-                      std::vector<double>& x, int N);
+bool chol_spd_inplace(std::vector<float>& A, int N);
+void chol_solve_lower(const std::vector<float>& L,
+                      const std::vector<float>& b,
+                      std::vector<float>& x, int N);
 
 } } // namespace top3d::mg
