+++ cpp/src/multigrid/preconditioner.cpp
@@ -35,14 +35,14 @@ void build_static_once(const Problem& pb, const MGPrecondConfig& cfg,
 Preconditioner make_diagonal_preconditioner_from_static(const Problem& pb,
 														  const MGHierarchy& H,
 														  const std::vector<std::vector<uint8_t>>& fixedMasks,
-														  const std::vector<double>& eleModulus,
+														  const std::vector<float>& eleModulus,
 														  const MGPrecondConfig& cfg) {
 	// 1) Build per-level diagonals
-	std::vector<std::vector<double>> diag;
+	std::vector<std::vector<float>> diag;
 	MG_BuildDiagonals(pb, H, fixedMasks, eleModulus, diag);
 
 	// 2) Build aggregated Ee at coarsest level and factorize
-	std::vector<double> Lcoarse; int Ncoarse = 0;
+	std::vector<float> Lcoarse; int Ncoarse = 0;
 	{
 		const auto& Lc = H.levels.back();
 		Ncoarse = 3*Lc.numNodes;
@@ -51,11 +51,11 @@ Preconditioner make_diagonal_preconditioner_from_static(const Problem& pb,
 			Ncoarse = 0; // diagonal fallback
 		} else {
 			// 1) Finest-grid modulus in structured order
-			std::vector<double> emFineFull;
+			std::vector<float> emFineFull;
 			EleMod_CompactToFull_Finest(pb, eleModulus, emFineFull);
 
 			// 2) Coarsest dense K via Galerkin triple products with fine-level BC mask
-			std::vector<double> Kc;
+			std::vector<float> Kc;
 			MG_AssembleCoarsestDenseK_Galerkin(pb, H, emFineFull, fixedMasks.front(), Kc);
 
 			// 3) Safety: impose coarsest-level BCs too
@@ -76,14 +76,14 @@ Preconditioner make_diagonal_preconditioner_from_static(const Problem& pb,
 	}
 
 	// 3) Return preconditioner closure (same as MG diagonal-only path)
-	return [H, diag, cfg, &pb, fixedMasks, Lcoarse, Ncoarse](const std::vector<double>& rFree, std::vector<double>& zFree) {
+	return [H, diag, cfg, &pb, fixedMasks, Lcoarse, Ncoarse](const std::vector<float>& rFree, std::vector<float>& zFree) {
 		const int n0_nodes = (pb.mesh.resX+1)*(pb.mesh.resY+1)*(pb.mesh.resZ+1);
 		const int n0_dofs  = 3*n0_nodes;
-		std::vector<double> r0(n0_dofs, 0.0);
+		std::vector<float> r0(n0_dofs, 0.0);
 		for (size_t i=0;i<pb.freeDofIndex.size();++i) r0[pb.freeDofIndex[i]] = rFree[i];
 
-		std::vector<std::vector<double>> rLv(H.levels.size());
-		std::vector<std::vector<double>> xLv(H.levels.size());
+		std::vector<std::vector<float>> rLv(H.levels.size());
+		std::vector<std::vector<float>> xLv(H.levels.size());
 		rLv[0] = r0; xLv[0].assign(n0_dofs, 0.0);
 
 		for (int i=0;i<n0_nodes;i++) {
@@ -98,12 +98,12 @@ Preconditioner make_diagonal_preconditioner_from_static(const Problem& pb,
 			for (int i=0;i<fn_nodes;i++) {
 				for (int c=0;c<3;c++) {
 					const int d = 3*i+c; if (fixedMasks[l][d]) continue;
-					xLv[l][d] += cfg.weight * rLv[l][d] / std::max(1e-30, D[d]);
+					xLv[l][d] += cfg.weight * rLv[l][d] / std::max(1e-30f, D[d]);
 				}
 			}
 			rLv[l+1].assign(3*cn_nodes, 0.0);
 			for (int c=0;c<3;c++) {
-				std::vector<double> rf(fn_nodes), rc;
+				std::vector<float> rf(fn_nodes), rc;
 				for (int i=0;i<fn_nodes;i++) rf[i] = rLv[l][3*i+c];
 				MG_Restrict_nodes(H.levels[l+1], H.levels[l], rf, rc);
 				for (int i=0;i<cn_nodes;i++) rLv[l+1][3*i+c] = rc[i];
@@ -123,15 +123,15 @@ Preconditioner make_diagonal_preconditioner_from_static(const Problem& pb,
 			xLv[Lidx].assign(3*cn_nodes, 0.0);
 			for (int i=0;i<cn_nodes;i++) for (int c=0;c<3;c++) {
 				const int d = 3*i+c; if (fixedMasks[Lidx][d]) { xLv[Lidx][d] = 0.0; continue; }
-				xLv[Lidx][d] = rLv[Lidx][d] / std::max(1e-30, D[d]);
+				xLv[Lidx][d] = rLv[Lidx][d] / std::max(1e-30f, D[d]);
 			}
 		}
 
 		for (int l=(int)H.levels.size()-2; l>=0; --l) {
 			const int fn_nodes = H.levels[l].numNodes;
-			std::vector<double> add(3*fn_nodes, 0.0);
+			std::vector<float> add(3*fn_nodes, 0.0);
 			for (int c=0;c<3;c++) {
-				std::vector<double> xc(H.levels[l+1].numNodes), xf;
+				std::vector<float> xc(H.levels[l+1].numNodes), xf;
 				for (int i=0;i<H.levels[l+1].numNodes;i++) xc[i] = xLv[l+1][3*i+c];
 				MG_Prolongate_nodes(H.levels[l+1], H.levels[l], xc, xf);
 				for (int i=0;i<fn_nodes;i++) add[3*i+c] = xf[i];
@@ -142,7 +142,7 @@ Preconditioner make_diagonal_preconditioner_from_static(const Problem& pb,
 			for (int i=0;i<fn_nodes;i++) for (int c=0;c<3;c++) if (fixedMasks[l][3*i+c]) xLv[l][3*i+c] = 0.0;
 			for (int i=0;i<fn_nodes;i++) for (int c=0;c<3;c++) {
 				int d = 3*i+c; if (fixedMasks[l][d]) continue;
-				xLv[l][d] += cfg.weight * rLv[l][d] / std::max(1e-30, D[d]);
+				xLv[l][d] += cfg.weight * rLv[l][d] / std::max(1e-30f, D[d]);
 			}
 		}
 		zFree.resize(rFree.size());
